Deprecated microsoft365defender pipeline for microsoftxdr pipeline
