+++ sigma/pipelines/microsoft365defender/__init__.py
@@ -1,6 +1,5 @@
-from .microsoft365defender import microsoft_365_defender_pipeline, microsoft_xdr_pipeline
+from .microsoft365defender import microsoft_365_defender_pipeline
 
 pipelines = {
-    "microsoft_365_defender_pipeline": microsoft_365_defender_pipeline,
-    "microsoft_xdr_pipeline": microsoft_xdr_pipeline,
+    "microsoft_365_defender_pipeline": microsoft_365_defender_pipeline,  # DEPRECATED: Use microsoft_xdr_pipeline instead.
 }
