+++ sigma/pipelines/microsoftxdr/__init__.py
@@ -1,0 +1,5 @@
+from .microsoftxdr import microsoft_xdr_pipeline
+
+pipelines = {
+    "microsoft_xdr_pipeline": microsoft_xdr_pipeline,
+}
