+++ sigma/pipelines/microsoftxdr/mappings.py
