+++ sigma/pipelines/microsoftxdr/schema.py
