+++ sigma/pipelines/microsoftxdr/tables.py
