+++ sigma/pipelines/microsoftxdr/transformations.py
